+++ systemtest/src/main/java/io/strimzi/systemtest/kafkaclients/clientproperties/ConsumerProperties.java
@@ -46,6 +46,7 @@ public class ConsumerProperties extends AbstractKafkaClientProperties<ConsumerPr
             return this;
         }
 
+        @SuppressWarnings("deprecation") // OffsetResetStrategy is deprecated
         public ConsumerPropertiesBuilder withAutoOffsetResetConfig(OffsetResetStrategy offsetResetConfig) {
 
             this.properties.setProperty(ConsumerConfig.AUTO_OFFSET_RESET_CONFIG, offsetResetConfig.name().toLowerCase(Locale.ENGLISH));
@@ -69,7 +70,7 @@ public class ConsumerProperties extends AbstractKafkaClientProperties<ConsumerPr
     }
 
     @Override
-    @SuppressWarnings({"unchecked"})
+    @SuppressWarnings({"unchecked", "deprecation"}) // OffsetResetStrategy is deprecated
     public ConsumerProperties.ConsumerPropertiesBuilder toBuilder(ConsumerProperties clientProperties) {
         ConsumerPropertiesBuilder builder = new ConsumerProperties.ConsumerPropertiesBuilder();